move StringBuf impl to resolve -Wstringop-overflow
